+++ src/main/webapp/js/controllers/IndexController.js
@@ -5,7 +5,7 @@
  * @constructor
  */
 
-app.controller("IndexController", function($scope, $http, $location, userService) {
+app.controller("IndexController", function($scope, $http, $location, $window, userService) {
 	
 	$scope.menu = {};
     $scope.handbook = [];
@@ -43,6 +43,10 @@ app.controller("IndexController", function($scope, $http, $location, userService
     $scope.addFavorite = function(uid, sub){
 		if (uid){
 			if (sub.favor){
+				var ret = $window.confirm('确认取消收藏?');  
+				if (!ret){
+					return;
+				}
 				$http.delete('/favor/'+uid+'/' + sub.id).success(function() {
 					sub.favor = false;
 		        });
