+++ src/main/webapp/js/controllers/ResultController.js
@@ -4,7 +4,7 @@
  * ResultController
  * @constructor
  */
-app.controller("ResultController", function($scope, $http, $location, userService) {
+app.controller("ResultController", function($scope, $http, $location, $window, userService) {
 	
 	$scope.menu = {};
 	$scope.results = {};
@@ -137,6 +137,10 @@ app.controller("ResultController", function($scope, $http, $location, userServic
     $scope.addFavorite = function(uid, sub){
 		if (uid){
 			if (sub.favor){
+				var ret = $window.confirm('确认取消收藏?');  
+				if (!ret){
+					return;
+				}
 				$http.delete('/favor/'+uid+'/' + sub.id).success(function() {
 					sub.favor = false;
 		        });
